+++ tests/testthat/metadata-3/logicalpath-9f16e8-DELETE.R
@@ -1,6 +1,6 @@
 structure(list(method = "DELETE", url = "/logicalpath?logical-path=%2FtempZone%2Fhome%2Frods%2Ftestthat%2Fsome_object.rds&no-trash=1&recursive=0", 
     status_code = 200L, headers = structure(list(Server = "nginx/1.23.1", 
-        Date = "Tue, 27 Jun 2023 05:57:28 GMT", `Content-Length` = "0", 
+        Date = "Thu, 29 Jun 2023 06:41:06 GMT", `Content-Length` = "0", 
         Connection = "keep-alive", `Access-Control-Allow-Origin` = "*", 
         `Access-Control-Allow-Headers` = "*", `Access-Control-Allow-Methods` = "AUTHORIZATION,ACCEPT,GET,POST,OPTIONS,PUT,DELETE"), class = "httr2_headers"), 
     body = raw(0)), class = "httr2_response")