+++ NEWS.md
@@ -1,3 +1,8 @@
+# rirods 0.1.1
+
+* Supply configuration file to correct user-specific configuration directory 
+with `rappdirs::user_config_dir()`.
+
 # rirods 0.1.0
 
 * Initial CRAN submission.